+++ app/layout.tsx
@@ -2,10 +2,10 @@ import "./globals.css";
 import type { Metadata } from "next";
 import { GeistSans } from "geist/font/sans";
 import { GeistMono } from "geist/font/mono";
-import { Nav } from "./components/ui/nav";
-import { Footer } from "./components/ui/footer";
+import { Nav } from "@/components/ui/nav";
+import { Footer } from "@/components/ui/footer";
 import { Toaster } from "@/components/ui/toaster";
-import { PersistentToast } from "./components/ui/persistent-toast";
+import { PersistentToast } from "@/components/ui/persistent-toast";
 import config from "@/config/config";
 
 export const metadata: Metadata = {